[SIDE-461] More TS Config, move /proto within src folder
