+++ src/controllerPlugin.js
@@ -1,6 +1,6 @@
 /** @module controllerPlugin */
 
-const services = require('../proto/ldk_grpc_pb');
+const services = require('./proto/ldk_grpc_pb');
 const { prepareLogging } = require('./logging');
 
 const BrokerGrpcServer = require('./brokerGrpcServer');
