+++ src/proto/stdio.proto
