+++ src/proto/stdio_grpc_pb.js
