+++ src/proto/stdio_pb.js
