+++ src/sensorGrpcServer.js
@@ -1,5 +1,5 @@
-const messages = require('../proto/ldk_pb');
-const services = require('../proto/ldk_grpc_pb');
+const messages = require('./proto/ldk_pb');
+const services = require('./proto/ldk_grpc_pb');
 
 const { categories } = require('./categories');
